Issue #17 : Enable deleting settings in snapshot wp-admin/post page.
Create link 'Remove setting,' which toggles to 'Restore setting' on click.
It also sets a hidden text field for each setting to be removed.
On saving, these appear in $_REQUEST[ 'customize_snapshot_remove_settings' ].
Filter post content  on 'content_save_pre,' when this is saved.
If $_REQUEST has settings to be removed, filter them out.
Also, add simple styling of this 'Remove setting' link.
It appears red when set to 'remove,' and blue when set to 'restore.'